Added All Functionallity in CalculatorActivity instead of Main. Bug Fixes
