+++ CalculatorApp/Resources/layout/activity_main.xml
@@ -79,7 +79,7 @@
             android:layout_height="0dp"
             app:layout_rowWeight="1"
             app:layout_columnWeight="1"
-            android:text="x"/>
+            android:text="*"/>
 
         <Button
             style="@style/calc_button"
